convert to own module. read toml once
