+++ snooty/specparser.py
@@ -4,7 +4,7 @@
 from __future__ import annotations
 
 import dataclasses
-from dataclasses import asdict, dataclass, field
+from dataclasses import dataclass, field
 from enum import Enum
 from typing import (
     Any,
@@ -260,47 +260,6 @@ class RstObject:
             name=self.name,
             rstobject=self,
         )
-
-
-@checked
-@dataclass
-class FacetDefinition:
-    name: str
-
-
-@checked
-@dataclass
-class TargetPlatformDefinition:
-    name: str
-    versions: List[FacetDefinition]
-
-
-@checked
-@dataclass
-class TaxonomySpec:
-
-    genres: List[FacetDefinition]
-    target_platforms: List[TargetPlatformDefinition]
-
-    @classmethod
-    def get_taxonomy(cls) -> TaxonomySpec:
-        path = util.PACKAGE_ROOT.joinpath("taxonomy.toml")
-        taxonomy = check_type(cls, tomli.loads(path.read_text(encoding="utf-8")))
-        return taxonomy
-
-    @classmethod
-    def validate_key_value_pairs(cls, facet_str_pairs: List[tuple[str, str]]) -> None:
-        taxonomy_ref = asdict(cls.get_taxonomy())
-        try:
-            while len(facet_str_pairs) > 0:
-                key, value = facet_str_pairs.pop()
-                list_values = taxonomy_ref[key] or []
-                found_values = [
-                    x for x in list_values if x == value or x["name"] == value
-                ]
-                taxonomy_ref = found_values[0]
-        except Exception as e:
-            raise KeyError(e)
 
 
 @checked
